rose/client/requirements.txt
requests==2.18.4
lxml==4.2.1
dnspython==1.15.0
xvfbwrapper==0.2.9
